Páginas com Segurança

+++ src/main/java/com/nintendo/buyback/model/Role.java
@@ -1,33 +1,35 @@
 package com.nintendo.buyback.model;
 
-import javax.persistence.*;
 
-/**
- * Created by avieira on 03/03/2017.
- */
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
 @Entity
 @Table(name = "role")
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name="role_id")
-    private int id;
+    private long id;
     @Column(name="role")
     private String role;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
-
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
-
     public String getRole() {
         return role;
     }
-
     public void setRole(String role) {
         this.role = role;
     }
+
+
 }
