+++ src/main/java/com/nintendo/buyback/repository/RoleRepository.java
@@ -1,15 +1,12 @@
 package com.nintendo.buyback.repository;
 
+
 import com.nintendo.buyback.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-/**
- * Created by avieira on 03/03/2017.
- */
-
 @Repository("roleRepository")
-public interface RoleRepository extends JpaRepository<Role, Integer> {
+public interface RoleRepository extends JpaRepository<Role, Long>{
     Role findByRole(String role);
 
 }
